requirements.txt
aiofiles==23.2.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cn2an==0.5.22
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.5
debugpy==1.8.0
decorator==5.1.1
einops==0.7.0
encodec==0.1.1
eng-to-ipa==0.0.2
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.1
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.44.3
fsspec==2023.10.0
future==0.18.3
gradio==3.41.2
gradio_client==0.5.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.19.4
idna==3.4
importlib-resources==6.1.1
inflect==7.0.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
langid==1.1.6
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
more-itertools==10.1.0
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai-whisper==20230124
opencv-python==4.8.1.78
orjson==3.9.10
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==4.0.0
proces==0.1.7
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pydub==0.25.1
pyyaml==6.0.1
Pygments==2.16.1
pyopenjtalk-prebuilt==0.3.0
pyparsing==3.1.1
pypinyin==0.49.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rpds-py==0.13.0
safetensors==0.4.0
scipy==1.11.3
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
stack-data==0.6.3
starlette==0.27.0
SudachiDict-core==20230927
SudachiPy==0.6.7
sympy==1.12
tokenizers==0.15.0
toolz==0.12.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.2
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.7
urllib3==2.1.0
uvicorn==0.24.0.post1
vocos==0.1.0
wcwidth==0.2.10
websockets==11.0.3
wget==3.2
